requirements.txt
setuptools==46.1.3
wheel==0.34.2
Django==3.0.2
aiogram==2.6.1
uvloop==0.14.0
ujson==1.35
vkbottle~=3.0
xmltodict==0.12.0
requests==2.25.1
yandex-geocoder==2.0.0
xmltodict==0.12.0
configparser==5.0.2
mysqlclient==2.0.3
six~=1.16.0
pymessenger~=0.0.7.0
pymorphy2~=0.9.1
google-cloud-language~=2.2.2